add overflow:hidden to .prefix class
I noticed that if width of the browser window was such that the content of the Prefix element was too big for the element, then it was overflowing onto the adjacent element (ie, the input), obscuring the input.  Adding overflow:hidden to the .prefix class should fix that.  It may also be needed for the .postfix class, but I have not yet used that or tried that.