Added the functionality to creat non-equilateral
With the advanced triangle you can make whatever type of triangles you want. You decide the length of each side, so you can make nonequilateral triangles. The rotation argument is meant to rotate the triangles so you can play as much as you want! See the example in action, to get a tiny clue of it's big potential