Fix for layout generator.
As of commit 24bd4e9 that introduced Slim layout to the generator, the
default ERB layout can't be generated anymore.

This commit fixes this by adding the correspondent option to the
generator and fixing the condition in the code.

Also, I made a minor cosmetic change in the USAGEi file for consistency.