Setting up a vertical rhythm for typography with compass
This uses the compass vertical rhythm mixin to setup the typography
settings in addition to modular scale. The fonts and line-heights are
set in pixels to make this work sensibly. When Compass changes to
support rems with a pixel fall back that can be used instead.

I've also added class based font sizes so that font sizes can be used
regardless of header level. It isn't all that common that H1 is always
the same size in all contexts and with the new HTML5 sematics it will be
pretty common for most headings in a hierarchy of content to be H1's.