Make Address Bar hiding conditional on no #Hash
If the user comes to a page via a deep link (eg, www.example.com/page/#anchor-on-page) we don't want the `scrollTo` to fire because it will prevent the page from staying at the anchor location (which is confusing, since the page would try to go to the anchor and then scroll up when the `scrollTo` fires). This commit remedies that via what should be a pretty standard, cross browser method (as far as I know). :)