fix custom form elements exclusion
I went looking for (and found!) the undocumented feature of the Custom Forms that if you add `data-customforms="disabled"` to an element in a custom form it will not be converted into the Foundation Custom version.  This is useful for instances when the developer wishes to use a particular form field in a different manner (eg, apply some other kind of custom form code to the field).  However, I discovered that the change I've made for this Pull Request is necessary for it to fully work.