Add beforeSlideTransition() to orbit
This is needed because the afterSlideTransition() fires after the
transition effect, while the buttons change before the slide change.

This should have almost no effect on people that don't need it. But,
people that need to make additional fades correspond to the trans-
ition have a hook to do so.