Sticky Top Bar
Using the top-bar with the class .sticky, it can be placed anywhere in
the document flow, for example, under a heading (when used with
.contain-to-grid). Once the page is scrolled to where the navigation
is, the class .fixed is added and it remains at the top of the page.
When scrolling back up, the class .fixed is removed once it hits it's
original position.