Adds more flexible options for setting the row width of the grid
Each responsive screen size from small, medium, large and xlarge allows
for a different width setting. There is also a global maximum width
setting which constrains the overall proportions of the grid (and is
useful for legacy browser support)

This also changes the way the `respondTo` mixin works slightly. It now
uses small, medium, large and xlarge no longer skipping large as a
choice and dropping the 'Screen' suffix to make it more concise.