FIX for grid-row($behavior: nest)
It goes wrong on line 116 in the .scss file. If you have a childrow within a parentrow, then the childrow is larger then the parantrow.
<div class="parant row">
  <div class="child row"></div>
</div>

Because of "box-sizing: border-box" it needs to add padding instead of subtracting.
The width is set to auto, but let say that the browser has given it a value of 500px internally.
Subtracting padding will lead to width - (-padding) = 500 - (-30px) = 530px.