Added 'active' class to clicked data-orbit-link
This allows the link to be styled differently once it is clicked. This
allows for better custom navigation for Orbit.

Each time a link is clicked any sibling data-orbit-link elements lose
the 'active' class. This code may need improvements though, because it
assumes that a parent 'ul' element encompasses all sibling links. If a
developer were to use something other than a 'ul' element, this code
wouldn't work and 'active' classes wouldn't be removed from sibling
elements.