'active' class on data-orbit-link is kept up to date with current Orbit slide
I know this code may be a bit clumsy so please improve it since I'm no
JavaScript guru :)

What I've done here is to update the 'active' class on a
data-orbit-link whenever a slide finishes animating. This improves the
possibility for custom Orbit navigation. For an example use case please
see http://new.damir.com.au/ and scroll down to the Testimonials
section. Now you can use the default Orbit previous and next arrows to
change the active slide and the appropriate data-orbit-link (photo of a
referee in this example) will be set to 'active'. In other words, the
custom navigation keeps up with the Orbit slider.