Custom Dropdown tab-compatibility
Before, custom dropdowns were confusing because they cannot be tabbed to like other inputs, but contain two links which can be. For two tab-presses, it appears the form is broken.

Now, users can tab to a dropdown, which opens it, and enables existing interactions. whether or not the dropdown is opened, users can also tab to the next input and the dropdown will close or stay closed.