Fix <script> displaying in inline-lists
I know it's not a very good code style, but sometimes `script` tag is used not only before closing `body` tag. For example when inserting social widgets code (it really better to keep it in the place you want the widget to be displayed).
This patch fixes displaying javascript code in inline lists.