Don't apply background and color style on links inside .has-form elements within top-bar-section.
Applying '.has-form' class on the list item adds extra padding for buttons.
Because of that padding when user points a cursor on the item but outside of the button the button's background will disapper.
This fix adds extra condition to check if elements doesn't have '.has-form' before applying the background style.