Vertically aligns left/right tooltips with their target
The previous calculation for the top offset of left/right tooltips
was based on the nub height rather than the height of the tooltip
itself.  This worked in some cases to approximately align the target
and tip vertically, but was wrong for most cases.

The new calculation takes in to account the offset of the target
from the top of the page, the height of the target, and the height
of the tip.  The result is that the tip is aligned vertically with
its target.