objPos rewrite to account for Multi-line : TIP-LEFT/TIP-RIGHT
When using DATA-WIDTH or any multiline tooltip, the tooltip is top aligned with the nub vertically centered and pointing at the wrong item.

Data-Width on Tooltip needs to be set first to calculate vertical middle point and tooltip height. Moved it outside of the objPos function. Updated tip-left and -right to move tooltip up by half the objects height, pointing num at target.