Add 'active' class to the link that triggered accordion content
In previous incarnations the link that opened accordion content be be targeted based on whether or not it was active. I think this was because the HTML structure was different before place the link within the element that the "active" class was applied to whereas now it precedes it.

This fix adds an "active" class to the link that triggered the accordion content, making the assumption it is the nearest preceding sibling.