adding $tabbar-link-color variable
The color of links in the tabbar defaults the the blue links which are hard to see on the grey background. Using this variable the link color defaults to white and the dependency on topbar-menu-link-color for the menu-icon is removed

This makes using icon fonts easier, so you can use different icons as fonts.

Signed-off-by: Maarten Ottens <m@maartenottens.com>