Fix orbit bullets not lining up horizontally.
Observed: orbit bullets are lined up vertically, with `display: block` style set instead of `display: inline-block`. This is because the styles of the `li` element of `.orbit-container` overwrites `li` element of `.orbit-bullets`.
Fix: Add additional specificity to `.orbit-container li` as `.orbit-container ul li` so that it applies only to the images of the Orbit, not to the bullets.