Fix reveal-data-id event binding
This patch fixes 2 issues. First, it removes a duplicate and probably
unintended removal of event handlers on elements with a `data-reveal-id`
attribute. Second, it fixes the event binding for elements with a
`data-reveal-id` attribute to also work with elements added after
foundation has been intialized (e.g. due to ajax calls).