Remove redundant $include-html-form-classes
I haven't used Switch before, but I had to hunt down why adding:

```scss
@import "foundation/components/buttons";
@import "foundation/components/forms";
```

was leaving out form/input styles. This was the culprit, and I don't see the purpose in defining this variable twice in the settings.

The main issue here, however, is the default form styles are married to the class-based styles in _form.scss. We should decouple them in order to import form styles without the overhead class styles (allowing for sweet-semantic HTML). I'll be adding another pull request respectively.