Include form styles by default
This is a proposed change to decouple form styles from classes. In order to include form styles currently, `$include-html-classes` must be `true`-- or one has to dig through the settings and switch on `$include-html-form-classes`.

The Foundation docs gives the user the impression that in order to enable form styles, all one must do is simply import the components. In summary: if we `@import` the form component, the styles should be enabled by default.