Refactored the reveal-modal-style mixin
The previous version of this mixin required that you define a ton of parameters each time you wanted to apply a style. When the mixin is first called in order to apply the default styles, those default styles are passed as parameters. Therefore in subsequent calls, you are only passing the particular attribute that you want to override.

This new version of the mixing seems to remove a small amount of duplicate styles that were not needed.

I made this a separate commit in case this reflector was not desired but you still wanted to accept the new radius, round and collapse classes.