Fixed Orbit height issues
This solves problems with instances of Orbit that do not have direct image children.

compute_dimensions is now also called on window.load() since this event get triggered after all assets on the page have been loaded.

I changed the children() search to use find(). We cannot always guarantee that there will be direct image children. So find will find all of the child images. I know that children() is faster than find() but we are talking about micro seconds here. I think that the added functionality is worth it.

Also the children are stored in a variable. This way find() is only called once and not twice like children() was.