Orbit: skip expensive progress computation (if disabled)
This adds the timer_show_progress_bar data option/setting, which skips
creation of the progress element.
Additionally the $orbit-timer-show-progress-bar SASS setting will
generate `display:none` for .orbit-progress.

If the progress element is not visisble (`display:none`) or does not
exist, the progress handling gets skipped.

This avoids the quite expensive animations, which are not visible
anyway, and saves us a lot of CPU cycles.

Fixes https://github.com/zurb/foundation/issues/4283