Fix outerHeight for jQuery 1.8.0 version
The parameter 'includeMargin' is set to false by default (https://api.jquery.com/outerHeight/). However due to a bug in jQuery 1.8.0 the outerHeight returns an object instead of an integer: http://bugs.jquery.com/ticket/12159

Adding 'false' fixes the equalizer for jQuery 1.8.0 and wouldn't do harm as it is set to false by default.