Fixed malformed Data URI
Removed a space from a URL in the form-select mixin that was causing
Ruby's URI.parse method to throw an error. Can't 100% guarantee that
spaces in data URIs are actually against spec, but my reading of RFC
2397 suggests they aren't, and it'll decode the same without the space
anyway.

Ran into the issue when using a gem (sinatra-assetpack) that attempts to
parse URLs in CSS to see if they can be "cache busted".