Enable dropdown bounds check so it can be guaranteed to be visible when open.
Dropdown will try to honor chosen align setting, unless the resulting position makes part of the dropdown content to be outside of chosen bounds. In that case, the alignment will be automatically update to make the whole dropdown visible in the chosen bounds.
Bounds setting "window" will display the dropdown in the currently visible document area, ensuring the user does not need to scroll to see the dropdown.
Bounds setting "document" may force the user to scroll to see the dropdown.
Turn this behaviour off by setting the "bounds" setting to anything else than "window" or "document".