Fixed broken regex from previous commit
There was a previous commit that broke many existing interchange instances because it was trying to add the ability to support commas inside the url path. In order to support backwards compatibility and this new edge case, more logic needed to be added. This should work in all cases now.

I also made sure that I left support in for defining new directives outside of the default 'replace'