Fix topbar dropdown unitless addition
I encountered a problem with the arrow padding while specifying the
topbar height in rems.

Previously, the topbar adds an extra 20 to the $topbar-link-padding which
is derived from $topbar-height. Adding an extra 20rem is very different
from adding the desired 20px.

This commit extracts that 20 unitless value into a variable so that the
units may change as needed.

Also considered having this commit set the height in rem so that it
adjusts with $rem-base, but I felt that it was specified in pixels for a
good reason.