Add an offset_by_height option to Magellan.
By default, when you click on a link on the Magellan expedition, it
offsets the destination by the height of the expedition. This behavior
isn't desired when the Magellan is on a sidebar and is vertically
oriented (what happens is that the destination seems to be offset for no
particular reason).

This commit adds an offset_by_height option to Magellan that when false
disables the aforementioned default behavior.