Removing 'resize' event when launching modal.
The window is not resized when a modal launches, therefore a resize event should not be triggered. There may be use cases where it would be helpful for a resize event to fire when a modal is launched, but it should be up to the individual developer to handle the event themselves.

Triggering a resize when a resize doesn't actually happen can mess up code that is looking for an actual resize event.