known issue "Abide - custom validation slow" #6547
I fixed some 4 spaces to 2 spaces and some spelling errors in the code.
nothing major there. But I also fixed the abide to be faster by lowering
the check time from 1second to 1/10th of a second. This resolves the
feeling of lag and does not seem to hinder any other function that I am
seeing. This does not resolve the 3 times execution as of yet. but as of
now does resolve the speed issue.