Include missing data_options
This includes all the missing data_options that are available on the individual Javascript components documentation pages.  I did not include any that are functions.  Note: some of the existing data options from this documentation are not on the individual components documentation pages.  Dropdowns, Reveal Modal, Tooltips are some examples of that.