Add additional check on el
Fixes a "Cannot read property 'nodeName' of undefined" error that people
seem to run into now and again.

I don't know what I did to get in a situation to get this error,
interchange works in other places on the same site fine, but adding this
little extra check fixed it.

Even if it was "my fault" somehow probably never a bad idea to check
what really matters here.