Global Keyboard Handler
Added 'handled' and 'unhandled' functions that are called if the event was or was not handled by the listener
Added handler to reveal
Added handler to tabs
Improved handler usage on range slider
Added handler to orbit, which now provides keyboard support for this feature. Even if the global handler will not be used, this should definitely be ported to the component!