Update foundation.topbar.js
Fixed an issue created by the intial-scroll of Google Chrome.

Description of the issue and the solution:
If you scroll down your foundation-website and use the sticky-tobbar and then you reload the page on Google Chrome, Chrome will directly jump to your last scroll position. Because start_offset has the default value of 0, the script won't remove the 'sticky' and 'fixed' classes. So I've set the default value of start_offset to false (still 0 in JS arithmetics) and on the initialization of the script, it searches for the current offset of the topbar and sets it to the start_offset variable, if the start_offset equals to false.