Complete rewrite based on feedback from Chris Oyler
This method of DOM mutation listening sets up individual listeners for
elements that require listening to in addition to the document body.
The document body will fire on any mutation event signaling to each
listening element to trigger its event handler and handle accordingly,
all with out looping over each element. In theory this can be expanded
to resize, scroll, and “closeme” triggers as well. As it sits now, this
will help any issues with elements that “change on the fly” and effect
other elements.