Issue #7557 Apply correct styling to hollow button group
Added logic to test for a hollow property within the button-fill variable. If it is present, applies the correct hollow styling to the buttons, even if there is a wrapping color class such as secondary or primary. Also fixes the transparent border being applied to non-stacked and stacked button groups.