Fix keyboard access for drilldown
Mainly I changed the focused elements to be the `a` tags rather than the `li` tags.
I also added anchors to the menu links in the docs, so it is possible to actually click/space/enter the links to open a new page (or at least change their window's location).
The fact that the links within the menus had only pure `#`es in their `href` did not cover the real-life menu purposes and made proper testing a bit hard.