Fix for detection of overfloating dropdown menus.
Because the selectors were not specific enough, the styles of .opens-right set by the JS in case of an overflowing sub-menu was overwritten by the parent's .open-left styles.
Added visual tests to check for this. The tests include two cases: one for a right aligned dropdown using `.top-bar-right` (which was fixed earlier already) and one that is positioned right because of content (where this fix can be tested).
Also added anchors to the menu's links to check for correct behavior of the links, e.g. when using keyboard.