Enhance keyboard util.
Enhanced keyboard util by adding a parameter that is passed to the `handled` function based on the evaluation of the function that was handled.
It is now easier to only do certain actions (e.g. calling `preventDefaul()`) in the `handled` handler by letting the other function (e.g. `next`, `up`, ...) return some value, which is directly passed into `handled` as the first parameter.
I also adjusted the calls to the keyboard util for some plugins to fix issues and prevent bad behavior.