Issue #7904 should be addressed by updating various ARIA related issues.
Removed aria-hidden on both the body and the overlay. Since the modals are children of both of those anything that has an aria-hidden=true will be ignored by the screen reader. This is why with a keyboard you can tab around int he modal but when you use a screen reader and you tab around you can't access the content within the modal.

I also enhanced the aria-labeledby by checking to see if the modal has a H Tag (H1,2,3,4,5,6) of sorts. If it finds a H Tag the first one found will be associated to the aria-labeledby and thus will result with that text announced to the a screen reader after opening a modal. if there is no header it will utilize the element (href, button, etc...) that opened the modal and associated itself to the aria-labeledby and thus will result with that text announced to the a screen reader after opening a modal.