Fix proportions of children in input-group
If I understand flex model correctly, there is no need to force the ‘.input-group-field‘ item to have certain width. Actually it breaks the design in some cases.

The point is, the flexbox container calculates the width of its items itself. If we set flex-grow to 1 for ‘.input-group-field‘ (which we did), the flexbox gives that item “width priority” over the rest, which is set 0 (default).