Update _top-bar.scss
Messing around with the editor in Chrome, it looks like it might currently be possible to adjust the `top-bar-title` class margin-right property to be a margin property. This is just a test. The scss/settings/_settings.scss file would also need to have the setting changed in there though. 

If this is an acceptable solution, I can look into it closer & make another pull request.