Fix expanded menu #8325
Fix https://github.com/zurb/foundation-sites/issues/8325.
See also: https://github.com/zurb/foundation-sites/pull/8779. Thanks to
@natewiebe13.

`.expanded` is a modifier for a horizontal menu, and should not be
applied on a vertical one.

Changes:
- Move the `.expanded` modifier after `.horizontal` and before
`.vertical`.
- Add the `.expanded` modifiers for each breakpoint with the
corresponding prefixes.

Other changes:
- Use `-zf-each-breakpoint` to generate responsive modifiers.

Note: The `.expanded` modifier should be applied only when the menu is
horizontal. For example, on a menu which is made horizontal on the
medium breakpoint, `.medium-expanded` should be used :
`.menu.vertical.medium-horizontal.medium-expanded`.