Fix size of row nested inside collapsed (flex-)grid row.
See: https://github.com/zurb/foundation-sites/issues/9557
Problem: Every row has a max width, so when nested, it cannot be
enlarged with negative margins to handle the paddings of its columns.

Fix: expand the size of all row nested in a non-expanded row.

Note: THIS IS A TEMPORARY FIX. Grids and flex-grids have a lot of bugs
like this one and should be refactorized. Please beat me if you see
this commit 6 months later and nothing changed.