Update for global and private variables, removed baseline
Inserted dropcaps-uppercase to global variable to let people change it. Used margin, font-size, line-height as a private variables assuming that this won't be changed. Also removed vertical-align:baseline as its useless when you are using float. Also now we are using the specific CSS property order as per foundation guidelines