Maths Magic for dropcaps to calculate font size and margin(top/bottom)
Changed lots of coding logic, first defined number of lines against the rest of the content and line height for the rest of the content. Then added these to a parameter for dropcaps mixin, then set the distance between ascent and top of the font to 0.15 and distance between baseline and bottom of the font to 0.18. Then multiplied line height and number of lines to a defined variable. Lastly had some maths magic for dropcaps to calculate font size, margin top and margin bottom and then inserting them to the main first-letter code