Changes to better dovetail with current Foundation design patterns.
Changed method of detecting disabled per
https://github.com/zurb/foundation-sites/pull/9812#discussion_r109806300

Altered docs to reflect current functionality of disable accordions,
documented below.

If an accordion is disabled, all motion methods (`up`, `down`, `toggle`)
get eaten, with the exception of `down` when combined with `firstTime`.
Added a `console.info()` message to indicate when these methods are
attempted on a disabled accordion.