Scripts in HTML Starter Template linked wrong
The links in the starter template were linked wrong, jquery.min.js doesn't exist in the CSS only download, but jquery.js does, same with the what-input script, it does not have a min version in the CSS download, only a version without min. The file foundation.min.js exists but it is in the vendor folder, not directly in the js folder. After fixing these three links in the starter template, everything I try is now working as expected. 

This starter template should link everything correctly so everything works. I actually was looking at other frameworks because I couldn't get all that I wanted to run working, this was why. I went back to my old Foundation code and everything is working perfectly. This template may just be old code from a previous version of foundation when the file structure was different.

This fix will help new people to this framework learn and stay with Foundation.