Slight tweak for combined row/column description
I had never fully understood what the row/column combination on a single element did exactly. Today I clicked on the "Edit in browser" link and found a different explanation in codepen. Key words here: "If you have just a single column, you can save some markup" -- A _single column_! This shed some light on the concept. 

Maybe it's just me, but the current text "If you need a full-width column to use as a container," had never been clear to me which width it was referring to. Also, the "save some markup" part also cleared up the purpose of this technique. So basically I'm just proposing changing the current description to be a little more like the Codepen description.