Add else statement inside of ifequal example
It is unclear that else statements are possible inside of {{#ifequal}} helpers to show HTML if the two values are not equal. Because there are {{#ifpage}} and {{#unlesspage}} as separate equal/not equal helpers, users may be confused that there is no equivalent "unlessequal" helper or "!" syntax.

I have successfully used the {{else}} Handlebars helper inside if the {{#ifequal}} Paini helper to achieve the desired "not equal" condition, but only thought to try this as a guess after a lot of unsuccessful searching through documentation on both Handlebars and Panini.