fix: wait for gulp tasks ends with the `finish` event
`pipe(done)` is a mistake and pipe does not work that way.
`on('end')` may not be triggered because `end` is an interval stream event that does not reflect the whole gulp stream operation. Only `finish` should be used to wait for the operation end.

See: https://stackoverflow.com/a/41003102/4317384