test: give time to IE to render elements before testing focus
For all the following tests testing the currently focused elements,
IE 9/10/11 may not focus the tested element because is is still
rendering. To prevent this, we wait after the component initialization.

Fixing this IE behavior in components would require to delay the focus
for all browsers then trigger a "ready" event.

See https://stackoverflow.com/a/2600261/4317384
See https://stackoverflow.com/a/36032615/4317384