fix: remove $-zf-size from non-responsive XY Grid classes generation
`$-zf-size` is only set within breakpoints. Breakpoints are not expected to be used around the `xy-vertical-grid-classes()` mixin.

Note: a more appropriate fix would be to rely on the `xy-cell()` mixin instead of `xy-cell-static()` to fully support explicit/implicit/null breakpoint option. However this may introduce breaking changes, so this will be kept for v6.6.0.

[This fix is compatible with v6.5]