feat: add `$output` to xy-cell mixin and deprecate `$output-gutter`
Add the `$output` list to the xy-cell mixin for a finer control of what should be output, and an explicit indication that each part may still have to be generated seperately.

`$gutter-output` still works the same way, and breaking change introduced in #11081 has been replaced with a deprecation notice. Migration notes and advices to correctly generate gutters will be included in the release notes of the future version