fix: make the "ignoreMousedisappear()" handler called before mouseenter
Do not debounce the received `mouseleave` event when a `mouseenter` is expected. This is not useful (we do not have to test the window focus in this case) and would make the `mouseenter` event handler called before the filtered `mouseleave` event handler.

See https://github.com/zurb/foundation-sites/pull/11418#issuecomment-407911537