fix: remove invalid "aria-multiselectable" on AccordionMenu #11437
As a menu, links inside the AccordionMenu plugin are only hidden for practical visual reasons. Opening a link "expand it" (make nested links visible), but no information (other than strictly visual) if added about the expanded link itself or the value it may represent.

For this reason, AccordionMenu's links cannot be considered as "selected" when opened but only "expanded", and the "aria-multiselectable" attribute should be removed.

Closes https://github.com/zurb/foundation-sites/issues/11437