docs: remove mentions of unsupported Y Grid from block grids section
Block Grid is not supported for Y Grid and will not be implemented. It would add too much code for a feature that is will not be used in most projects.

Plus, Block Grid classes are not namespaced for the horizontal or vertical grid. Adding block grid classes for the vertical grid would lead of inconstencies or backward incompatibility and increased specificity.

Closes https://github.com/zurb/foundation-sites/issues/11463