fix: keep the original event datas in Touch tap/swipe events (v6.5)
Keep the `touchstart`/`touchmove`/`touchend` datas (like `target`) for the events that are manually triggered: `tap` and `swipe`.

This resolve a bug in "Reveal" when a `touchend` event inside the reveal modal is transformed to a `tap` event on the overlay because the original `target` is lost when triggered the event on the overlay.

This is a patch of f016a9f (#11430) for v6.5.