chore: temporary disable tree-shaking #11586
Despite the previous attempts to make Foundation compatible with tree shaking (#11445, #11497 and #11508), Foundation is used with side-effects (on jQuery) and cannot be considered as "without side-effects" for tree-shaking purposes.

This commit remove the "sideEffects: false" label and thus disable tree-shaking support, until the Foundation API and usage is changed to not implicitely rely on the global jQuery object.

Closes https://github.com/zurb/foundation-sites/issues/11586

Related to:
- https://github.com/zurb/foundation-sites/pull/11508
- https://github.com/zurb/foundation-sites/pull/11497
- https://github.com/zurb/foundation-sites/pull/11445