remove TAB/SHIFT_TAB from script to allow browser to manage tabbing
While the tab key moves focus appropriately in the dom order within this component. The focus is not visually discernible when the subnav is opened. Since the method to move back up within the dom tree is present via arrow keys and the link at the top of the subnav, I agree with the solution that was proposed in https://github.com/foundation/foundation-sites/issues/11641
Broad information about keyboard best practice can be found here: https://www.w3.org/TR/wai-aria-practices/#kbd_generalnav