fix: error when switching ResponsiveAccordionTabs from tabs to accordion
When switching from tabs to accordion, the tabs instance gets destroyed.
Because destroying tabs and recalculating tabs height in  _setHeight method
rely on the foundatuion mediaquery event, both are executed.
Even if tabs instance got destroyed before.

This behaviour leads to an error in the _setHeight of tabs, because tabs markup is
already destroyed and $tabContent is null.

To prevent this, $tabcontent gets tested to be not falsy.

Closes #12106